refactor: move postgres specific code into submodule
